+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## TBD
+
+* Amend secondary instance URL to bugsnag.smartbear.com [#122](https://github.com/bugsnag/bugsnag-flutter-performance/pull/122)
+
 ## 1.7.0 (2025-09-22)
 
 ### Changes